requirements.txt
ccxt==3.1.49
pyyaml==6.0.1
numpy==1.24.4
pandas==2.0.3
matplotlib==3.7.2
schedule==1.2.0
requests==2.31.0 